Further refactoring of package layout.
